Feat(UI): Implement Repeater (UI Template Engine)
- Implemented UI_NODE_LIST logic in ui_def.c for array iteration.
- Updated codegen.py to support pointers to pointers and size_t.
- Reflected MathGraph and MathNode for binding.
- Added count_source to UiDef for explicit array counting.
- Added test_repeater to ui_tests.c to verify functionality.

+++ src/domains/math_model/math_graph.c
@@ -18,7 +18,7 @@ static void free_node(MathNode* node) {
 
 void math_graph_dispose(MathGraph* graph) {
     if (!graph) return;
-    for (size_t i = 0; i < graph->node_count; ++i) {
+    for (int i = 0; i < graph->node_count; ++i) {
         free_node(graph->nodes[i]);
     }
     free(graph->nodes);
